Count recently assigned bugs
In order to improve balance across team members when assigning new bugs, show also for each developer the number of bugs assigned in the past 21 days (moving window, symbolically equivalent to one sprint).
Backports generated by Prow bot are excluded from the count.

The recently-assigned count requires n new queries to the JIRA server, with n = # of developers in the team. In order to skip this costly evaluation, the new command line argument "-q" or "--quick" should be used.

Improved output when a query to JIRA server fails: now a single dot is printed instead of the full error, which we wouldn't be able to fix on our side anyway.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>